Closes https://github.com/aimakerspace/PeekingDuck/issues/519

Changes include:
- Change underlying mosaic algorithm to use 2x `cv2.resize()`
  - A downsize to (`mosaic_level`, `mosaic_level`) followed by an upsize to the original size
- Change both `mosaic_bbox` and `mosaic` to private
- Convert `mosaic` to an instance method so it can access `self.mosaic_level` directly without have to be passed in as an argument
- Check that `bbox` coordinates are within bounds

There are some differences in the mosaiced output between the current implementation and PR implementation:
| Current | PR | Remarks |
|---------|----|---------|
| ![jim_curr](https://user-images.githubusercontent.com/56420072/142155866-1debe75e-df12-4998-8deb-f3b02e490770.gif) | ![jim_fast](https://user-images.githubusercontent.com/56420072/142155879-03f7a67a-5798-49ed-98cc-1ee4e388cc91.gif) | 1280 x 720 video with `mosaic_level=7`<br> Output may appear jittery as it is more sensitive to source pixel color |
| ![michael_curr](https://user-images.githubusercontent.com/56420072/142155580-88647248-cac7-45d0-b480-063d23c14a17.gif) | ![michae_fast](https://user-images.githubusercontent.com/56420072/142155589-b63b6346-bf2b-4961-8cad-9991debc53e6.gif) | 1280 x 720 video with `mosaic_level=17`<br> Jitteriness is less noticeable with higher `mosaic_level` |
| ![multiple_curr](https://user-images.githubusercontent.com/56420072/142156028-0ea37e7b-fe77-452a-80fe-77e14b9853c6.gif) | ![multiple_fast](https://user-images.githubusercontent.com/56420072/142156078-3390e89c-5856-4011-b19a-5bd5f5a9fdd5.gif) | 1280 x 720 video with `mosaic_level=7`<br> Less noticeable with smaller bounding boxes too |

While a robust comparison was not performed, the PR implementation provides an improvement in FPS performance for rows 2 and 3 in the above table:
- 23.94 vs 25.26 FPS
- 19.99 vs 21.49 FPS

**NOTE**: Similar to the current implementation, the PR implementation does not check that an appropriate `mosaic_level` is used. When `mosaic_level` is large or approximately same as bbox size, the output is not "properly" mosaiced, i.e., the person's privacy is not actually protected.