@ongtw 2021-11-19
Code looks good to go.

Just two minor comments:
1. Past tense of mosaic (verb) is spelled `mosaicked`. Can update documentation accordingly.
2. In above PR examples, would be instructive to mention the `mosaic_level` used, to give users an idea of what are good mosaic level values.