@liyier90 2021-11-19
Fixed typo and added video resolution with `mosaic_level` used for the examples in the opening comment.